+++ CHANGES.rst
@@ -3,7 +3,7 @@
 Version 2.1.2
 -------------
 
-Unreleased
+Released 2022-04-28
 
 -   Fix type annotation for ``json.loads``, it accepts str or bytes.
     :issue:`4519`